PERF - WIP Test integration of queues. Test performance reporting completeness. Perf-metrics unit testing in-place, already.

+++ src/main/java/performance/Performance.java
@@ -14,5 +14,5 @@ public interface Performance {
     double getReliabilityPercentage();
     double getTotalProfit();
     double getProfitFactor();
-    int getTradeCount();
+    long getTradeCount();
 }
